% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/url_exists.R
\name{cbind_na}
\alias{cbind_na}
\title{function for merging r objects (chars, int, data.frame, ...) of different sizes}
\usage{
cbind_na(..., first = TRUE)
}
\arguments{
\item{...}{R objects}

\item{first}{ysorting na values}
}
\value{
a data frame
}
\description{
function for merging r objects (chars, int, data.frame, ...) of different sizes
}
\examples{
library(afnews)
cbind_na(a=1:2, b=letters[1:5], c=c('a',1,3))
}
