% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ml_news.R
\name{ml_news_contents}
\alias{ml_news_contents}
\title{Main functions of collecting news from Mali}
\usage{
ml_news_contents(start_d, end_d, press = "all")
}
\arguments{
\item{start_d}{yyyy-mm-dd start date of collection (oldest date, eg: yesterday)}

\item{end_d}{yyyy-mm-dd collection end date (newest date, eg: today)}

\item{press}{specify the name of a press.}
}
\value{
a table containing the news (title + content + date publisher + press) collected
}
\description{
Function for collecting news (title + content) from the media in Mali.
}
\examples{
library(afnews)
dt_news <- ml_news_contents(start_d="2020-10-20", end_d="2020-10-28", press='all')
head(dt_news)
}
