% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/simdata.R
\name{mvrnormR}
\alias{mvrnormR}
\title{Simulate Multivariate Normal Distribution}
\usage{
mvrnormR(n, p, rho)
}
\arguments{
\item{n}{sample size (integer)}

\item{p}{number of variables (integer)}

\item{rho}{correlation between variables (between 0 and 1)}
}
\value{
A nxp matrix of simulated data
}
\description{
Custom function to simulate multivariate normal distribution
}
\examples{
dat <- mvrnormR(n = 100, p = 10, rho = 0.8)
heatmap(cor(dat))
}
\references{
\url{https://gallery.rcpp.org/articles/simulate-multivariate-normal/}
}
