% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/sn_news.R
\name{sn_news_contents}
\alias{sn_news_contents}
\title{Main functions of collecting information on Senegal}
\usage{
sn_news_contents(start_d, end_d, press = "all")
}
\arguments{
\item{start_d}{yyyy-mm-dd start date of collection (oldest date, eg: yesterday)}

\item{end_d}{yyyy-mm-dd collection end date (newest date, eg: today)}

\item{press}{specify the name of a press.}
}
\value{
a table containing the news (title + content + date publisher + press) collected
}
\description{
Function for collecting news (title + content) from the media in Burkina Faso.
}
\examples{
library(afnews)
dt_news <- sn_news_contents(start_d="2020-10-28", end_d="2020-11-02", press='all')
head(dt_news)
}
