% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/url_exists.R
\name{url_exists}
\alias{url_exists}
\title{Function to check if URL exists}
\usage{
url_exists(x, non_2xx_return_value = FALSE, quiet = FALSE, ...)
}
\arguments{
\item{x}{a single URL}

\item{non_2xx_return_value}{what to do if the site exists but the
HTTP status code is not in the \verb{2xx} range. Default is to return \code{FALSE}.}

\item{quiet}{if not \code{FALSE}, then every time the \code{non_2xx_return_value} condition
arises a warning message will be displayed. Default is \code{FALSE}.}

\item{...}{other params (\code{timeout()} would be a good one) passed directly
to \code{httr::HEAD()} and/or \code{httr::GET()}}
}
\value{
return TRUE or FALSE
}
\description{
Function to check if URL exists
}
\examples{
urls <- c('http://www.amazon.com', 'http://this.isafakelink.biz', 'https://stackoverflow.com')
base::sapply(urls,url_exists)
}
